hf-space/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
transformers>=4.49.0
accelerate>=0.26.0
huggingface_hub>=0.27.0,<0.28.0
qwen-vl-utils==0.0.8
gradio>=5.0.0
Pillow>=10.0.0
